t0061: Work around a line endings issue with newer versions of cat on MSYS
While msysgit uses cat 2.0, mingwGitDevEnv uses cat 5.97, currently. A
change part of the latter version [1] makes cat use the "same rules as
other programs to decide whether to use binary I/O". As a result, cat 5.97
behaves like e.g. awk with respect to which line endings are used if
stdout is redirected to a file, and those are DOS line endings. Ignore
line endings when comparing the test result to work around this.

[1] http://git.savannah.gnu.org/gitweb/?p=coreutils.git;a=commit;h=8770c00ef45e3c6c4dd3d5ce1e55a6fb1adb13dc

Signed-off-by: Thomas Braun <thomas.braun@virtuell-zuhause.de>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>